[refactor] Refactor TableCache to interface with ZkTableCache implementation
